the module of sdk-spi-aws is separated to multiple modules
