+++ capa-spi-aws-mesh/src/main/java/group/rxcloud/capa/spi/aws/mesh/config/AwsClientProviderLoader.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package group.rxcloud.capa.spi.aws.config;
+package group.rxcloud.capa.spi.aws.mesh.config;
 
 import group.rxcloud.cloudruntimes.utils.TypeRef;
 import software.amazon.awssdk.core.SdkClient;
