+++ capa-spi-aws-mesh/src/main/java/group/rxcloud/capa/spi/aws/mesh/package-info.java
@@ -15,4 +15,4 @@
  * limitations under the License.
  */
 
-package group.rxcloud.capa.spi.aws.configuration;
+package group.rxcloud.capa.spi.aws.mesh;
