+++ capa-spi-aws/src/main/resources/capa-component.properties
@@ -1,0 +1,2 @@
+group.rxcloud.capa.component.http.CapaHttp=group.rxcloud.capa.spi.aws.mesh.http.AwsCapaHttp
+group.rxcloud.capa.spi.config.CapaSpiOptionsLoader=group.rxcloud.capa.spi.aws.mesh.config.AwsSpiOptionsLoader
