--- sdk-spi-aws/src/main/resources/capa-component.properties
@@ -1,2 +1,0 @@
-group.rxcloud.capa.component.http.CapaHttp=group.rxcloud.capa.spi.aws.http.AwsCapaHttp
-group.rxcloud.capa.spi.config.CapaSpiOptionsLoader=group.rxcloud.capa.spi.aws.config.AwsSpiOptionsLoader
